+++ hazelcast/src/test/java/com/hazelcast/test/AbstractHazelcastClassRunner.java
@@ -17,9 +17,6 @@
 package com.hazelcast.test;
 
 import com.hazelcast.cache.jsr.JsrTestUtil;
-import com.hazelcast.client.HazelcastClient;
-import com.hazelcast.core.Hazelcast;
-import com.hazelcast.core.HazelcastInstance;
 import com.hazelcast.internal.util.ConcurrencyUtil;
 import com.hazelcast.test.annotation.Repeat;
 import com.hazelcast.test.bounce.BounceMemberRule;
@@ -45,7 +42,6 @@ import java.lang.management.ThreadInfo;
 import java.lang.management.ThreadMXBean;
 import java.lang.reflect.Method;
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
@@ -53,8 +49,6 @@ import java.util.Set;
 import java.util.concurrent.ForkJoinPool;
 import java.util.concurrent.TimeUnit;
 
-import static com.hazelcast.cache.jsr.JsrTestUtil.clearCachingProviderRegistry;
-import static com.hazelcast.cache.jsr.JsrTestUtil.getCachingProviderRegistrySize;
 import static com.hazelcast.internal.util.EmptyStatement.ignore;
 import static com.hazelcast.test.TestEnvironment.isRunningCompatibilityTest;
 import static java.lang.Integer.getInteger;
@@ -302,37 +296,7 @@ public abstract class AbstractHazelcastClassRunner extends AbstractParameterized
     @Override
     protected Statement withAfterClasses(Statement statement) {
         final Statement originalStatement = super.withAfterClasses(statement);
-        return new Statement() {
-            @Override
-            public void evaluate() throws Throwable {
-                originalStatement.evaluate();
-
-                // check for running Hazelcast instances
-                Set<HazelcastInstance> instances = Hazelcast.getAllHazelcastInstances();
-                if (!instances.isEmpty()) {
-                    String message = "Instances haven't been shut down: " + instances;
-                    Hazelcast.shutdownAll();
-                    throw new IllegalStateException(message);
-                }
-                Collection<HazelcastInstance> clientInstances = HazelcastClient.getAllHazelcastClients();
-                if (!clientInstances.isEmpty()) {
-                    String message = "Client instances haven't been shut down: " + clientInstances;
-                    HazelcastClient.shutdownAll();
-                    throw new IllegalStateException(message);
-                }
-
-                // check for leftover JMX beans
-                JmxLeakHelper.checkJmxBeans();
-
-                // check for leftover CachingProvider instances
-                int registrySize = getCachingProviderRegistrySize();
-                if (registrySize > 0) {
-                    clearCachingProviderRegistry();
-                    throw new IllegalStateException(registrySize + " CachingProviders are not cleaned up."
-                            + " Please use JsrTestUtil.cleanup() in your test!");
-                }
-            }
-        };
+        return new AfterClassesStatement(originalStatement);
     }
 
     private String generateThreadDump() {
